refactor(stl-deserializer): extracted binary reader * adde stubs for output type selection option
